requirements.txt
# python
black
python-dotenv

# data
ibis-framework[duckdb,bigquery,deltalake] @ git+https://github.com/ibis-project/ibis

# viz
plotly
streamlit

# ops
dagster
dagster-webserver

# ai
openai
marvin

# cloud
azure-identity
azure-storage-blob
azure-storage-file-datalake --pre
